separate undo/redo into it's own plugin
