+++ src/pages/c/universe.astro
@@ -1,6 +1,7 @@
 ---
 import { TextEditor } from '../../components/universe/index'
 import { KeyboardShortcutsPlugin } from '../../components/universe/plugins/shortcuts'
+import { UniverseUndoPlugin } from '../../components/universe/plugins/undo'
 import Layout from '../../layouts/Layout.astro'
 ---
 
@@ -45,6 +46,7 @@ import Layout from '../../layouts/Layout.astro'
         <!-- @ts-ignore attributes -->
         <TextEditor client:only="lit" wrap>
           <KeyboardShortcutsPlugin client:only="lit" />
+          <UniverseUndoPlugin client:only="lit" />
         </TextEditor>
       </div>
     </div>
